refactor rmpf_worker_setup
